+++ src/components/beta/gux-table-beta/gux-table-constants.ts
@@ -1,0 +1,5 @@
+export interface IColumnResizeState {
+  resizableColumn: HTMLElement;
+  resizableColumnInitialWidth: number;
+  columnResizeMouseStartX: number;
+}